add Haskell implementations of the Chinese Remainder Theorem
I thought that I might need them for Day 8 Part 2. The sequences of nodes visited by following the path is periodic, but you may end at a node ending in Z multiple times during each period, so there may be a large number of possible systems of congruences to solve, so I didn't pursue this approach.